Android refactor: Move all contentUri utilities to the new class ContentUri
